+++ src/main/java/SpringJPA/Model/Administrator.java
@@ -1,4 +1,4 @@
-package main.models;
+package SpringJPA.Model;
 
 public class Administrator extends User{
