+++ src/main/java/SpringJPA/Model/Constraint.java
@@ -1,4 +1,4 @@
-package main.models;
+package SpringJPA.Model;
 
 import java.util.ArrayList;
 import java.util.List;
