+++ src/main/java/SpringJPA/Model/User.java
@@ -1,4 +1,4 @@
-package main.models;
+package SpringJPA.Model;
 
 public class User {
