+++ src/main/java/SpringJPA/WebpageApplication.java
@@ -1,4 +1,4 @@
-package sysc4806.project.webpage;
+package SpringJPA;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
